move Crafty to jmplib
