+++ src/main/java/xyz/jpenilla/wanderingtrades/util/StoredPlayers.java
@@ -14,7 +14,12 @@ import xyz.jpenilla.wanderingtrades.WanderingTrades;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.*;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
